Write tiffs as 32 bit integers instead of floats
Tiff files cannot be opened in image viewers if the pixel values are in floats. I want to still look at tiff files in image viewers from time to time.
Might consider changing to 16 bit in the future, as the original data is 14 bit anyways.

+++ src/tri_ct_tools/image/writer.py
@@ -30,5 +30,5 @@ def array_to_tif(img: np.ndarray, output_folder: Path, filename: str):
     output_folder.mkdir(parents=True, exist_ok=True)
     output_file = output_folder / filename
     output_tif = output_file.with_suffix(".tif")
-    tifffile.imwrite(output_tif, img)
+    tifffile.imwrite(output_tif, img.astype(np.int32))
     return output_tif
